+++ LuckyoneUI/Core/Defaults.lua
@@ -240,6 +240,7 @@ Private.Defaults = {
 		},
 		skins = {
 			Auctionator = false,
+			BigWigs = false,
 			BugSack = false,
 			DejaClassicStats = false,
 			LeatrixPlus = false,